Begin fleshing out a transaction based access log mechanism for Lusca.
One thing that annoyed me in Squid/Lusca is that it's difficult to
get fine-grained debugging grouped by the "transaction". It also
doesn't provide useful debugging about the forwarding decisions being made
(again, no xid) or the retry info (again, no xid.)

This access_log function takes an optional proxy_request pointer which
contains the transaction id.

Right now the "what" and "who" are just strings, specifically for me
to debug what's going on. Once the design is finalised, this will be
turned into a set of registered enums which will make it easier to
enable/disable debugging as needed. (Remember though, this needs to be
done _before_ the function arguments are evaluated, or debugging will
incur a significant overhead whether it's enabled or not.)