Begin moving the dns request handling out of lusca.c and into dns.c.
Previously, the dns code calls either dns_dispatch_requests() or
dns_dispatch_error(). The dns entry code had a list of proxy_request
structs linked to it.

This makes it difficult to use for things other than proxy requests,
and it also makes the evthr modifications more difficult then they
should be. Now:

* there's a new struct which encapsulates the dns requests, with
  a shared cache entry;
* that's walked in the two above functions, but it uses "arg" now.

The next step is to add the callback function pointers and status.